Ported code examples to 64 bit selfie

+++ manuscript/code/function.c
@@ -1,13 +1,13 @@
-int x;
+uint64_t x;
 
-int f(int x) {
+uint64_t f(uint64_t x) {
   while (x > 0)
     x = x - 1;
 
   return x;
 }
 
-int main() {
+uint64_t main() {
   x = 0;
 
   x = x + 1;
